requirements.txt
pip==25.0
Flask==3.1.0
Flask_Cors==5.0.0
openai==1.60.1
PyPDF2==3.0.1
pytest==8.3.4
python-dotenv==1.0.1
qdrant_client==1.13.2
sentence_transformers==3.3.1
supabase==2.12.0
